Complete rewrite of the render method, adding support for 8 and 16 bits depth (untested)
the render method now use a renderInternal templated function to do code factorization, like many openfx-misc plugin.
the render method now correctly support RGBA, RGB and Alpha components.
the render method now support depth of 8 and 16 bits, but it's not tested because I don't know how to do it with natron.
the plugin now tell to the host it support 8 and 16 bits.