Deleting some input:
Delete R, G, B and A checkbox because it's useless.
Delete mask sourceClip because I have finally decide to not implement the associated functionnality. Equivalent functionnality can be achieved with the "Merge" node of natron.
Silent an unused parameter warning on the "context" variable. I can't delete it because it comes from an interface.